Replacing a broken reward and a milestone that couldn't be completed due to the broken reward
Specifically, Phase 3 Milestone 6 (Unbounded) had a blank reward. I believe this is because an item or reward broke, maybe due to a bad id. Now instead of the "Memory of Bones" plans, 3 x Memory Fragments and a Carrier AI Fragment it only rewards a Carrier AI Fragment. Because you can't get the plans, Phase 4 Milestone 6 became impossible to complete, so I replaced it with some other simple milestone so you could complete the chain and finish the expedition.