Making steps scrollable
We've gotten lots of expeditions lately so the left column is getting unwieldy. Making it scrollable to make sure notes are visible. Not super happy about this since I used css grids, but if it's possible with flexbox I just didn't have the patience.

I'm leaving the "sticky" download button code even though it's non-functional due to the new scrolling.